+++ secrets/ed25519_test.go
@@ -6,4 +6,5 @@ func TestEd25519(t *testing.T) {
 	// TODO: GenerateEd25519Keys
 	// TODO: LoadEd25519PrivateKey
 	// TODO: LoadEd25519PublicKey
+	// TODO: Ed25519SignMessage
 }